feat: improve nextjs to use sdk hooks
