+++ packages/devnext/src/components/WalletActions.tsx
@@ -1,34 +1,58 @@
-import { CommunicationLayerPreference, ConnectionStatus, EventType, ServiceStatus } from '@metamask/sdk-communication-layer';
-import Head from 'next/head';
-import styles from 'src/styles/Home.module.css';
-import { MetaMaskSDK } from '@metamask/sdk';
-import { useEffect, useState } from 'react';
-import { MetaMaskInpageProvider } from "@metamask/providers";
-import { ModalLoader } from '@metamask/sdk-install-modal-web';
-import Link from 'next/link';
-import { useSDK } from '../../../sdk-react/src/MetaMaskHooks';
+import { useSignMessage, useSignTypedData } from "@metamask/sdk-react";
 
-let _initialized = false;
-export default function Home() {
-  const { sdk, connected, connecting, status: serviceStatus, account, chainId: chain, error } = useSDK();
-  const [response, setResponse] = useState<unknown>("");
+// All properties on a domain are optional
+const domain = {
+  name: 'Ether Mail',
+  version: '1',
+  chainId: 1,
+  verifyingContract: '0xCcCCccccCCCCcCCCCCCcCcCccCcCCCcCcccccccC',
+} as const
 
-  const connect = () => {
-    if (!window.ethereum) {
-      throw new Error(`invalid ethereum provider`);
-    }
-    window.ethereum
-      .request({
-        method: "eth_requestAccounts",
-        params: [],
-      })
-      .then((accounts) => {
-        if (accounts) {
-          console.debug(`connect:: accounts result`, accounts);
-        }
-      })
-      .catch((e) => console.log("request accounts ERR", e));
-  };
+// The named list of all type definitions
+const types = {
+  Person: [
+    { name: 'name', type: 'string' },
+    { name: 'wallet', type: 'address' },
+  ],
+  Mail: [
+    { name: 'from', type: 'Person' },
+    { name: 'to', type: 'Person' },
+    { name: 'contents', type: 'string' },
+  ],
+}
+
+const message = {
+  from: {
+    name: 'Cow',
+    wallet: '0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826',
+  },
+  to: {
+    name: 'Bob',
+    wallet: '0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB',
+  },
+  contents: 'Hello, Bob!',
+} as const
+
+const buttonClass = 'bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded';
+
+export const WalletActions = () => {
+  const { data, isError, isLoading, isSuccess, signTypedData } =
+    useSignTypedData({
+      domain,
+      types,
+      primaryType: 'Mail',
+      message,
+    });
+
+  const {
+    data: signData,
+    isError: isSignError,
+    isLoading: isSignLoading,
+    isSuccess: isSignSuccess,
+    signMessage,
+  } = useSignMessage({
+    message: 'gm wagmi frens',
+  });
 
   const eth_signTypedData_v4 = async () => {
     const msgParams = JSON.stringify({
@@ -114,73 +138,27 @@ export default function Home() {
       console.debug(`ethRequest ${method}`, JSON.stringify(params, null, 4))
       console.debug(`sign params`, params);
       const resp = await window.ethereum?.request({ method, params });
-      setResponse(resp);
+      console.debug(`response:`, resp);
     } catch (e) {
       console.log(e);
     }
   };
 
-  const personalSign = async () => {
-
-  }
-
-  const terminate = () => {
-    sdk?.terminate();
-  }
-
-  return (
-    <>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <header>
-        <Link href={'uikit'}>UI Kit demo</Link>
-      </header>
-      <main className={styles.main}>
-        <p>
-          Connection Status: <strong>{serviceStatus?.connectionStatus}</strong>
-        </p>
-        {serviceStatus?.connectionStatus === ConnectionStatus.WAITING &&
-          <div>
-            Waiting for Metamask to link the connection...
-          </div>
-        }
-        <p>ChannelId: {serviceStatus?.channelConfig?.channelId}</p>
-        <p>{`Expiration: ${serviceStatus?.channelConfig?.validUntil ?? ''}`}</p>
-
-        <div>
-          {`Connected: ${connected}`}
-        </div>
-        {connected ? <div>
-          <button style={{ padding: 10, margin: 10 }} onClick={connect}>
-            Request Accounts
-          </button>
-
-          <button style={{ padding: 10, margin: 10 }} onClick={eth_signTypedData_v4}>
-            eth_signTypedData_v4
-          </button>
-
-        </div> : <button style={{ padding: 10, margin: 10 }} onClick={connect}>
-          Connect
-        </button>}
-
-        <button style={{ padding: 10, margin: 10, backgroundColor: 'red' }} onClick={terminate} >
-          Terminate
-        </button>
-
-        <div>
-          <>
-            {chain && `Connected chain: ${chain}`}
-            <p></p>
-            {account && `Connected account: ${account}`}
-            <p></p>
-            {response && `Last request response: ${response}`}
-          </>
-        </div>
-      </main>
-    </>
-  );
+  return (<div style={{ marginTop: 20 }} className={'w-full'}>
+    <button className={buttonClass} onClick={eth_signTypedData_v4}>
+      Direct CALL
+    </button>
+    <hr/>
+    <button className={buttonClass} disabled={isLoading} onClick={() => signTypedData()}>
+      signTypedData
+    </button>
+    {isSuccess && <div className='text-ellipsis overflow-hidden'>Signature: {data}</div>}
+    {isError && <div>Error signing message</div>}
+    <p></p>
+    <button className={buttonClass} disabled={isSignLoading} onClick={() => signMessage()}>
+      sign
+    </button>
+    {isSignSuccess && <div className='text-ellipsis overflow-hidden'>Signature: {signData}</div>}
+    {isSignError && <div>Error signing message</div>}
+  </div>)
 }
