+++ packages/devnext/src/pages/_app.tsx
@@ -1,32 +1,17 @@
 // eslint-disable-next-line import/no-unassigned-import
-import '../styles/globals.css';
+import { MetaMaskProvider } from '@metamask/sdk-react';
 import type { AppProps } from 'next/app';
-import { configureChains, createConfig, mainnet, MetaMaskProvider, WagmiConfig } from '@metamask/sdk-react';
-import { useEffect, useState } from 'react';
-import { MetaMaskSDKOptions } from '@metamask/sdk';
-import { getDefaultProvider } from 'ethers'
-import { publicProvider } from '@wagmi/core/providers/public';
-
-const { chains, publicClient, webSocketPublicClient } = configureChains(
-  [mainnet],
-  [publicProvider()],
-)
-
-const serverConfig = createConfig({
-  autoConnect: true,
-  publicClient,
-})
+import '../styles/globals.css';
 
 export default function App({ Component, pageProps }: AppProps) {
-  const [sdkOptions, setSDKOptions] = useState<MetaMaskSDKOptions>();
-  const [clientSide, setClientSide] = useState(false);
-
-  useEffect(() => {
-    setSDKOptions({
+  return (
+    <MetaMaskProvider debug={false} sdkOptions={{
       // communicationServerUrl: 'http://192.168.50.114:4000',
       logging: {
         developerMode: false,
         sdk: false,
+        remoteLayer: false,
+        serviceLayer: false,
         plaintext: true,
       },
       autoConnect: {
@@ -39,18 +24,6 @@ export default function App({ Component, pageProps }: AppProps) {
         name: 'DevNext',
         url: 'http://devnext.fakeurl.com',
       }
-    })
-    setClientSide(true);
-  }, []);
-
-  return (
-    <>
-      {
-        clientSide ?
-          <MetaMaskProvider sdkOptions={sdkOptions}><Component {...pageProps} /></MetaMaskProvider> :
-          (
-            <WagmiConfig config={serverConfig}><Component {...pageProps} /></WagmiConfig>
-          )
-      }
-    </>);
+    }}><Component {...pageProps} /></MetaMaskProvider>
+  );
 }
