+++ packages/sdk/src/config.ts
@@ -15,4 +15,5 @@ export const STORAGE_PATH = '.sdk-comm';
 
 export const RPC_METHODS = {
   METAMASK_GETPROVIDERSTATE: 'metamask_getProviderState',
+  ETH_REQUESTACCOUNTS: 'eth_requestAccounts',
 };
